Added New Option of {player} Variable
Added permission: The command now requires the `globaltoast.use` permission to be executed.

Improved error handling: The code now checks for various error conditions, such as empty arguments, invalid target player, and console usage.

Added target player option: You can specify a target player as the last argument in the command. If a target player is specified, the toast will only be visible to that player. If no target player is specified, the toast will be visible to the executor.

Added "everyone" option: If the target player option is set to "everyone" (case-insensitive), the global toast will be sent to everyone in the server. This is achieved by cloning the ToastRequestPacket and sending it to each online player.

Updated the target player handling: If the target player option is set to "everyone", the code iterates over each online player to send the cloned packet. Otherwise, it sends the packet to the specified target player.